move app into container component
