+++ src/Containers/AppContainer.js
@@ -1,0 +1,4 @@
+import { connect } from 'react-redux';
+import App from '../App';
+
+export default connect(null)(App);
